+++ src/flowTypes.js
@@ -53,6 +53,7 @@ export type User = {
   alias: string,
   wins: ?number,
   losses: ?number,
+  valid: boolean,
 };
 
 export type AuthenticatedUser = User & {
@@ -81,6 +82,7 @@ export const defaultUser: User = {
   alias: '',
   wins: null,
   losses: null,
+  valid: true,
 };
 
 // See: https://wiki.guildwars2.com/wiki/API:2/guild/:id/log
@@ -103,6 +105,14 @@ type Log = {
   coins: number,
   action: 'queued' | 'cancelled' | 'completed' | 'sped_up',
   upgrade_id: number,
+};
+
+export type Token = {
+  accountName: string,
+  primary: boolean,
+  permissions: string,
+  valid: boolean,
+  token: string,
 };
 
 export type Guild = {